api/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
psycopg2-binary>=2.9.9
python-dotenv==1.0.1
openai==1.12.0
ffmpeg-python==0.2.0
moviepy==1.0.3
funasr>=1.1.0
modelscope>=1.9.0
python-multipart==0.0.9
httpx
pydantic==2.6.1
pydantic-settings>=2.1.0,<3
loguru==0.7.2
setuptools>=65.0.0
wheel
urllib3<2.0.0
opencv-python>=4.8.0
numpy>=1.24.0
celery>=5.3.0
redis>=5.0.0
